add jsdocs
