+++ src/control-plane/billing/index.ts
@@ -1,5 +1,5 @@
 // Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
 // SPDX-License-Identifier: Apache-2.0
 
-export * from './billing-template';
+export * from './billing-provider';
 export * from './billing-interface';
