+++ src/control-plane/ingestor-aggregator/ingestor-aggregator-interface.ts
@@ -4,9 +4,23 @@
 import * as dynamodb from 'aws-cdk-lib/aws-dynamodb';
 import * as lambda from 'aws-cdk-lib/aws-lambda';
 
-// todo: come up with a better name
+/**
+ * Encapsulates the list of properties for a IDataIngestorAggregator.
+ */
 export interface IDataIngestorAggregator {
-  readonly dataRepository: dynamodb.Table;
+  /**
+   * The table containing the aggregated data.
+   */
+  readonly dataRepository: dynamodb.ITable;
+
+  /**
+   * The function responsible for aggregating the raw data coming in
+   * via the dataIngestor.
+   */
   readonly dataAggregator: lambda.IFunction;
+
+  /**
+   * The ingestor responsible for accepting and storing the incoming data.
+   */
   readonly dataIngestorName: string;
 }
